fix typography; restructure skywire-cli config, visor, and hv subcommands. Remove hv srvpk subcommand. Limit hv skychat command to linux. Allow skywire-cli visor start as non root
