+++ cmd/skywire-cli/commands/hv/hvskychat.go
@@ -1,20 +1,22 @@
-package hvui
+//go:build linux
+// +build linux
+package hv
 
 import (
-	"github.com/skycoin/skycoin/src/util/logging"
 	"github.com/spf13/cobra"
 	"github.com/toqueteos/webbrowser"
 )
 
-var logger = logging.MustGetLogger("skywire-cli:launch-browser")
-
+func init() {
+	RootCmd.AddCommand(chatCmd)
+}
 // RootCmd contains commands that interact with the skywire-visor
-var RootCmd = &cobra.Command{
-	Use:   "ui",
-	Short: "hypervisor UI",
+var chatCmd = &cobra.Command{
+	Use:   "skychat",
+	Short: "skychat UI",
 	Run: func(_ *cobra.Command, _ []string) {
 		//TODO: get the actual port from config instead of using default value here
-		if err := webbrowser.Open("http://127.0.0.1:8000/"); err != nil {
+		if err := webbrowser.Open("http://127.0.0.1:8001/"); err != nil {
 			logger.Fatal("Failed to open hypervisor UI in browser:", err)
 		}
 	},
