+++ cmd/skywire-cli/commands/hv/hvui.go
@@ -1,20 +1,20 @@
-package hvskychat
+package hv
 
 import (
-	"github.com/skycoin/skycoin/src/util/logging"
 	"github.com/spf13/cobra"
 	"github.com/toqueteos/webbrowser"
 )
 
-var logger = logging.MustGetLogger("skywire-cli")
-
+func init() {
+	RootCmd.AddCommand(hvuiCmd)
+}
 // RootCmd contains commands that interact with the skywire-visor
-var RootCmd = &cobra.Command{
-	Use:   "skychat",
-	Short: "skychat UI",
+var hvuiCmd = &cobra.Command{
+	Use:   "ui",
+	Short: "hypervisor UI",
 	Run: func(_ *cobra.Command, _ []string) {
 		//TODO: get the actual port from config instead of using default value here
-		if err := webbrowser.Open("http://127.0.0.1:8001/"); err != nil {
+		if err := webbrowser.Open("http://127.0.0.1:8000/"); err != nil {
 			logger.Fatal("Failed to open hypervisor UI in browser:", err)
 		}
 	},
