+++ cmd/skywire-cli/commands/hv/hvvpn.go
@@ -1,4 +1,4 @@
-package hvvpn
+package hv
 
 import (
 	"fmt"
@@ -10,16 +10,23 @@ import (
 	"github.com/skycoin/skywire/pkg/visor/visorconfig"
 )
 
-var path string
-var pkg bool
-
 func init() {
-	RootCmd.AddCommand(vpnUICmd)
+	RootCmd.AddCommand(vpnCmd)
+	vpnCmd.PersistentFlags().StringVarP(&rpcAddr, "rpc", "", "localhost:3435", "RPC server address")
+	vpnCmd.AddCommand(
+		vpnUICmd,
+		vpnURLCmd,
+		vpnListCmd,
+	)
 	vpnUICmd.Flags().StringVarP(&path, "input", "i", "", "read from specified config file")
 	vpnUICmd.Flags().BoolVarP(&pkg, "pkg", "p", false, "read from /opt/skywire/skywire.json")
-	RootCmd.AddCommand(vpnURLCmd)
 	vpnURLCmd.Flags().StringVarP(&path, "input", "i", "", "read from specified config file")
 	vpnURLCmd.Flags().BoolVarP(&pkg, "pkg", "p", false, "read from /opt/skywire/skywire.json")
+}
+
+var vpnCmd = &cobra.Command{
+	Use:   "vpn",
+	Short: "vpn interface",
 }
 
 var vpnUICmd = &cobra.Command{
@@ -73,5 +80,16 @@ var vpnURLCmd = &cobra.Command{
 			url = fmt.Sprintf("http://127.0.0.1:8000/#/vpn/%s/", overview.PubKey.Hex())
 		}
 		fmt.Println(url)
+	},
+}
+
+
+var vpnListCmd = &cobra.Command{
+	Use:   "list",
+	Short: "List public VPN servers",
+	Run: func(_ *cobra.Command, _ []string) {
+		client := rpcClient()
+		servers := client.VPNServers()
+		fmt.Println(servers)
 	},
 }
