+++ cmd/skywire-cli/commands/visor/app.go
@@ -1,4 +1,4 @@
-package vapps
+package clivisor
 
 import (
 	"fmt"
@@ -15,7 +15,8 @@ import (
 
 func init() {
 	cobra.EnableCommandSorting = false
-	RootCmd.AddCommand(
+	RootCmd.AddCommand(appCmd)
+	appCmd.AddCommand(
 		lsAppsCmd,
 		startAppCmd,
 		stopAppCmd,
@@ -24,17 +25,20 @@ func init() {
 	)
 }
 
+var appCmd = &cobra.Command{
+	Use:   "app",
+	Short: "App settings",
+}
+
 var lsAppsCmd = &cobra.Command{
 	Use:   "ls",
-	Short: "list apps",
+	Short: "List apps",
 	Run: func(_ *cobra.Command, _ []string) {
 		states, err := rpcClient().Apps()
 		internal.Catch(err)
-
 		w := tabwriter.NewWriter(os.Stdout, 0, 0, 5, ' ', tabwriter.TabIndent)
 		_, err = fmt.Fprintln(w, "app\tports\tauto_start\tstatus")
 		internal.Catch(err)
-
 		for _, state := range states {
 			status := "stopped"
 			if state.Status == appserver.AppStatusRunning {
@@ -52,7 +56,7 @@ var lsAppsCmd = &cobra.Command{
 
 var startAppCmd = &cobra.Command{
 	Use:   "start <name>",
-	Short: "launch app",
+	Short: "Launch app",
 	Args:  cobra.MinimumNArgs(1),
 	Run: func(_ *cobra.Command, args []string) {
 		internal.Catch(rpcClient().StartApp(args[0]))
@@ -62,7 +66,7 @@ var startAppCmd = &cobra.Command{
 
 var stopAppCmd = &cobra.Command{
 	Use:   "stop <name>",
-	Short: "halt app",
+	Short: "Halt app",
 	Args:  cobra.MinimumNArgs(1),
 	Run: func(_ *cobra.Command, args []string) {
 		internal.Catch(rpcClient().StopApp(args[0]))
@@ -72,7 +76,7 @@ var stopAppCmd = &cobra.Command{
 
 var setAppAutostartCmd = &cobra.Command{
 	Use:   "autostart <name> (on|off)",
-	Short: "set autostart flag for app",
+	Short: "Autostart app",
 	Args:  cobra.MinimumNArgs(2),
 	Run: func(_ *cobra.Command, args []string) {
 		var autostart bool
@@ -91,11 +95,10 @@ var setAppAutostartCmd = &cobra.Command{
 
 var appLogsSinceCmd = &cobra.Command{
 	Use:   "log <name> <timestamp>",
-	Short: "logs from app since RFC3339Nano-formated timestamp.\n                    \"beginning\" is a special timestamp to fetch all the logs",
+	Short: "Logs from app since RFC3339Nano-formated timestamp.\n                    \"beginning\" is a special timestamp to fetch all the logs",
 	Args:  cobra.MinimumNArgs(2),
 	Run: func(_ *cobra.Command, args []string) {
 		var t time.Time
-
 		if args[1] == "beginning" {
 			t = time.Unix(0, 0)
 		} else {
