+++ cmd/skywire-cli/commands/visor/root.go
@@ -1,4 +1,4 @@
-package visor
+package clivisor
 
 import (
 	"net"
@@ -7,9 +7,6 @@ import (
 	"github.com/skycoin/skycoin/src/util/logging"
 	"github.com/spf13/cobra"
 
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/visor/vapps"
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/visor/vroute"
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/visor/vtp"
 	"github.com/skycoin/skywire/pkg/visor"
 )
 
@@ -18,9 +15,6 @@ var logger = logging.MustGetLogger("skywire-cli")
 var rpcAddr string
 
 func init() {
-	RootCmd.AddCommand(vapps.RootCmd)
-	RootCmd.AddCommand(vroute.RootCmd)
-	RootCmd.AddCommand(vtp.RootCmd)
 	RootCmd.PersistentFlags().StringVarP(&rpcAddr, "rpc", "", "localhost:3435", "RPC server address")
 }
