+++ cmd/skywire-cli/commands/visor/shutdown.go
@@ -1,4 +1,4 @@
-package visor
+package clivisor
 
 import (
 	"fmt"
@@ -7,14 +7,12 @@ import (
 )
 
 func init() {
-	RootCmd.AddCommand(
-		shutdownCmd,
-	)
+	RootCmd.AddCommand(shutdownCmd)
 }
 
 var shutdownCmd = &cobra.Command{
 	Use:   "halt",
-	Short: "stop a running visor",
+	Short: "Stop a running visor",
 	Run: func(_ *cobra.Command, args []string) {
 		rpcClient().Shutdown() //nolint
 		fmt.Println("Visor was shut down")
