+++ cmd/skywire-cli/commands/visor/start.go
@@ -1,4 +1,4 @@
-package visor
+package clivisor
 
 import (
 	"os/user"
@@ -8,33 +8,34 @@ import (
 )
 
 var sourcerun bool
+var root bool
 
 func init() {
 	usrLvl, err := user.Current()
 	if err != nil {
 		panic(err)
 	}
-	var root bool
 	if usrLvl.Username == "root" {
 		root = true
 	}
-	//enforce root permissions for VPN operability
-	if root {
-		RootCmd.AddCommand(
-			startCmd,
-		)
-		startCmd.Flags().BoolVarP(&sourcerun, "src", "s", false, "'go run' external commands from the skywire sources")
-	}
+	RootCmd.AddCommand(startCmd)
+	startCmd.Flags().BoolVarP(&sourcerun, "src", "s", false, "'go run' external commands from the skywire sources")
 }
+
 
 var startCmd = &cobra.Command{
 	Use:   "start",
-	Short: "start a visor",
+	Short: "Start a visor",
 	Run: func(_ *cobra.Command, args []string) {
 		var err error
 		if !sourcerun {
-			//if skywire is installed as a command from a package, we can use the -p flag here
-			_, err = script.Exec(`skywire-visor -p`).Stdout()
+			if root {
+				//if skywire is installed as a command from a package, we can use the -p flag here
+				_, err = script.Exec(`skywire-visor -p`).Stdout()
+			} else {
+				//if the config exists in the userspace and this command was not run as root
+				_, err = script.Exec(`skywire-visor -u`).Stdout()
+			}
 		} else {
 			_, err = script.Exec(`bash -c 'go run cmd/skywire-visor/skywire-visor.go'`).Stdout()
 		}
