--- cmd/skywire-cli/commands/visor/vtp/transport_discovery.go
@@ -1,89 +1,0 @@
-package vtp
-
-import (
-	"errors"
-	"fmt"
-	"os"
-	"text/tabwriter"
-
-	"github.com/google/uuid"
-	"github.com/spf13/cobra"
-
-	"github.com/skycoin/skywire-utilities/pkg/cipher"
-	"github.com/skycoin/skywire/cmd/skywire-cli/internal"
-	"github.com/skycoin/skywire/pkg/transport"
-)
-
-func init() {
-	RootCmd.AddCommand(discTpCmd)
-}
-
-var (
-	tpID transportID
-	tpPK cipher.PubKey
-)
-
-func init() {
-	discTpCmd.Flags().Var(&tpID, "id", "if specified, obtains a single transport of given ID")
-	discTpCmd.Flags().Var(&tpPK, "pk", "if specified, obtains transports associated with given public key")
-}
-
-var discTpCmd = &cobra.Command{
-	Use:   "disc (--id=<transport-id> | --pk=<edge-public-key>)",
-	Short: "discover transport(s) by ID or public key",
-	Args: func(_ *cobra.Command, _ []string) error {
-		var (
-			nilID = uuid.UUID(tpID) == (uuid.UUID{})
-			nilPK = tpPK.Null()
-		)
-		if nilID && nilPK {
-			return errors.New("must specify --id flag or --pk flag")
-		}
-		if !nilID && !nilPK {
-			return errors.New("cannot specify --id and --pk flag")
-		}
-		return nil
-	},
-	Run: func(_ *cobra.Command, _ []string) {
-
-		if rc := rpcClient(); tpPK.Null() {
-			entry, err := rc.DiscoverTransportByID(uuid.UUID(tpID))
-			internal.Catch(err)
-			printTransportEntries(entry)
-		} else {
-			entries, err := rc.DiscoverTransportsByPK(tpPK)
-			internal.Catch(err)
-			printTransportEntries(entries...)
-		}
-	},
-}
-
-func printTransportEntries(entries ...*transport.Entry) {
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 5, ' ', tabwriter.TabIndent)
-	_, err := fmt.Fprintln(w, "id\ttype\tedge1\tedge2")
-	internal.Catch(err)
-	for _, e := range entries {
-		_, err := fmt.Fprintf(w, "%s\t%s\t%s\t%s\n",
-			e.ID, e.Type, e.Edges[0], e.Edges[1])
-		internal.Catch(err)
-	}
-	internal.Catch(w.Flush())
-}
-
-type transportID uuid.UUID
-
-// String implements pflag.Value
-func (t transportID) String() string { return uuid.UUID(t).String() }
-
-// Type implements pflag.Value
-func (transportID) Type() string { return "transportID" }
-
-// Set implements pflag.Value
-func (t *transportID) Set(s string) error {
-	tID, err := uuid.Parse(s)
-	if err != nil {
-		return err
-	}
-	*t = transportID(tID)
-	return nil
-}
